requirements.txt
Pillow
ImageHash
tqdm
